When the user completes a combo, consume the last keypress/release of the combo.
For example if the combo is ALT,N (in that order) then we consume the N keypress/release. If the user instead presses N,ALT then we consume the ALT keypress/release.

We only do this for the LAST key because it'll look like keys are broken if we speculatively consume all keys in the combo. In the ALT+N case, consuming all keys would make it look like the N key doesn't work or is delayed, since we'd keep grabbing it until it's apparent that a combo isn't being performed, like when a different key outside the combo is pressed. On top of that, there'd also be correctly handling the case when N is being held down. And then finally, there's the fact that multiple combos could be activated by the same key, like the ALT+P/ALT+N defaults.

For now this consume only applies when sending input to remote clients, since we don't grab server-local input (yet), so server->client switches will still see the last key emitted at the server.